backend/pyproject.toml
[project]
name = "ai-career-transition-planner-backend"
version = "0.1.0"
description = "AI Career Transition Planner - FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.7.0",
    "PyMuPDF>=1.25.0",
    "spacy>=3.8.0",
    "google-genai>=1.0.0",
    "groq>=0.33.0",
    "sentence-transformers>=3.3.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
